Rewrote core.
This commit adds an 'api' function that can call the Geni API and do
both reads and writes. Also added are 'read' and 'write' convenience
functions. In paticular, write automatically sets the request method to
POST so that you don't have to explicitly pass it to do a write.

This makes things much more simplistic. The Geni API is simple and
consistent enough to make it possible to just have one function for
interacting with it. Some APIs tend to be inconsistent in that they
require wildly different input in different APIs. For example, the
Github API has some methods that require JSON input while others require
simple form or query params. Geni has nothing like that.